Charging schedule and chart is now stored in the logs volume on each run.
