+++ src/HWChargeOptimizer/HWChargeOptimizer/Program.cs
@@ -5,10 +5,9 @@ using Microsoft.Extensions.Logging;
 using Serilog;
 using Serilog.Events;
 using HWChargeOptimizer.Configuration;
-using HWChargeOptimizer.Homewizard;
+using HWChargeOptimizer.HomeWizard;
 using HWChargeOptimizer.Reporter;
 using HWChargeOptimizer.Zonneplan;
-using HomewizardAuthentication = HWChargeOptimizer.Homewizard.HomewizardAuthentication;
 using ZonneplanAuthentication = HWChargeOptimizer.Zonneplan.ZonneplanAuthentication;
 
 using Microsoft.Extensions.Configuration;
@@ -56,7 +55,7 @@ builder.Services.AddHttpClient("NoSslValidation")
 
 builder.Services.AddHttpClient();
 builder.Services.AddTransient<IZonneplanAuthentication, ZonneplanAuthentication>();
-builder.Services.AddSingleton<IHomewizardAuthentication, HomewizardAuthentication>();
+builder.Services.AddSingleton<IHomeWizardAuthentication, HomeWizardAuthentication>();
 builder.Services.AddSingleton<IHomeWizardBatteryController, HomeWizardBatteryController>();
 builder.Services.AddTransient<IZonneplanTariffReader, ZonneplanTariffReader>();
 builder.Services.AddTransient<ChargeScheduleReporter>();
@@ -76,17 +75,10 @@ if (args.Length > 0)
             await reporter.RunAsync();
             break;
         
-        case "--chart":
-        case "-c":
-            var reportChartHost = builder.Build();
-            var reporterChart = reportChartHost.Services.GetRequiredService<ChargeScheduleReporter>();
-            await reporterChart.RunAsync(true);
-            break;
-        
         case "--service":
         case "-s":
             builder.Services.AddHostedService<ZonneplanScheduleService>();
-            builder.Services.AddHostedService<HomewizardScheduleService>();
+            builder.Services.AddHostedService<HomeWizardScheduleService>();
     
             var serviceHost = builder.Build();
